flask-backend/requirements.txt
Flask>=2.1.3
opencv-python>=4.6.0.66
numpy>=1.23.1
Pillow>=9.2.0
marshmallow>=3.17.0
Werkzeug>=2.2.1
requests>=2.28.1